+++ common/src/main/java/io/piano/android/common/DeviceIdProvider.kt
@@ -1,0 +1,11 @@
+package io.piano.android.common
+
+public class DeviceIdProvider(
+    commonPrefsStorage: CommonPrefsStorage,
+) {
+    public val deviceId: String = commonPrefsStorage.deviceId
+
+    public companion object {
+        public const val DEVICE_ID_HEADER: String = "Device-Id"
+    }
+}